Get Category from Virustotal

Returns category of content from 6 major services including: BitDefender, Dr. Web, Alexa (DMOZ), Google, Websense, and Trendmicro. Not all services will have categories for all the domains. When the categories are not returned for a particular domain, we return a NA.

virustotal_cat(domain = NULL, apikey = NULL)

Arguments

domain
domain name

apikey
virustotal API key

Value

data.frame with 7 columns: domain, bitdefender, dr_web, alexa, google, websense, trendmicro

Details

Get the API Access Key from http://www.virustotal.com/. Either pass the API Key to the function or set the environmental variable: VirustotalToken. Environment variables persist within a R session.

References

https://www.virustotal.com/en/documentation/public-api/

Examples

if (FALSE) {
virustotal_cat("http://www.google.com")
}